requirements.txt
boto==2.43.0
click==6.6
cssselect==1.0.0
cssutils==1.0.1
Django==1.10
django-appconf==1.0.1
django-compressor==2.1
django-ses-backend==0.1.1
djmail==0.11.0
futures==3.0.3
lxml==3.6.4
premailer==2.9.6
psycopg2==2.6.1
python-dotenv==0.6.1
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
